Separada la logida de JavaScript en un hook personalizado: useNewArticleLogic.js, ademas se configuro el archivo peticionesNewArticle.js porque daba falla al realizar las solicitudes post y patch
